Change version to string, make semantic

+++ lang/src/org/partiql/lang/planner/transforms/AstToLogicalVisitorTransform.kt
@@ -16,7 +16,7 @@ internal fun PartiqlAst.Statement.toLogicalPlan(): PartiqlLogical.Plan =
     PartiqlLogical.build {
         plan(
             AstToLogicalVisitorTransform.transformStatement(this@toLogicalPlan),
-            version = PLAN_VERSION_NUMBER.toLong()
+            version = PLAN_VERSION_NUMBER
         )
     }
